Don't use @ prefix in attribution in changelog.
I've decided that using the @ symbol as a username prefix outside of
GitHub is confusing, as it appears to be a Twitter handle, which the
referenced user may not control. However, when copying the changelog
text into "Releases" in the GitHub UI, I should use the @ symbol because
it creates a link to the referenced user.