Avoid window state from having invalid comic number.
If we can't load a comic from the internet or cache, we should set the
ComicNumber to 1, which will always be a valid comic number. This will
avoid issues where a user does a first launch without internet. When
they relaunch with internet, their state will put them at comic #1
instead of comic #0 (which doesn't exist).